ENH: dataframe now supports partial string selection (#1278)
* ENH: dataframe now supports partial string selection

* Split indexing.py

* Remove unnecessary statement
